refactor(e2e): move ensureE2eHttpbin into cluster-helpers
Follow overlays convention: cluster-facing helpers live beside
patchHttpbin rather than inline in the failswitch spec.
