+++ workspaces/orchestrator/e2e-tests/tests/specs/orchestrator-workflow-core.tests.ts
@@ -1,14 +1,11 @@
-import { writeFileSync } from "node:fs";
-import { tmpdir } from "node:os";
-import { join } from "node:path";
 import { test } from "@red-hat-developer-hub/e2e-test-utils/test";
 import { OrchestratorPage } from "@red-hat-developer-hub/e2e-test-utils/pages";
 import { OrchestratorPO } from "../support/pages/orchestrator-po.js";
 import {
+  ensureE2eHttpbin,
   patchHttpbin,
   restartAndWait,
   cleanupAfterTest,
-  runOc,
 } from "../support/utils/test-helpers.js";
 
 type EnsureDataIndexOrSkip = (
@@ -178,64 +175,4 @@ export function registerOrchestratorCoreWorkflowTests(
       await orchestrator.validateWorkflowAllRuns();
     });
   });
-}
-
-/** Minimal in-cluster /get mock so recovery does not depend on public httpbin.org. */
-function ensureE2eHttpbin(ns: string): void {
-  const manifest = `
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: e2e-httpbin
-  namespace: ${ns}
-spec:
-  replicas: 1
-  selector:
-    matchLabels:
-      app: e2e-httpbin
-  template:
-    metadata:
-      labels:
-        app: e2e-httpbin
-    spec:
-      containers:
-        - name: httpbin
-          image: registry.access.redhat.com/ubi9/python-311@sha256:a0bdb55576fc5b8d6704279307817828ef027e1065533ceba133fe9516003a6c
-          command:
-            - python3
-            - -c
-            - |
-              from http.server import HTTPServer, BaseHTTPRequestHandler
-              class H(BaseHTTPRequestHandler):
-                def do_GET(self):
-                  b=b'{"args":{},"headers":{},"origin":"e2e","url":"http://e2e-httpbin/get"}'
-                  self.send_response(200); self.send_header("Content-Type","application/json"); self.send_header("Content-Length",str(len(b))); self.end_headers(); self.wfile.write(b)
-                def log_message(self,*_): pass
-              HTTPServer(("0.0.0.0",8080),H).serve_forever()
-          ports:
-            - containerPort: 8080
-          readinessProbe:
-            httpGet:
-              path: /get
-              port: 8080
----
-apiVersion: v1
-kind: Service
-metadata:
-  name: e2e-httpbin
-  namespace: ${ns}
-spec:
-  selector:
-    app: e2e-httpbin
-  ports:
-    - port: 80
-      targetPort: 8080
-`;
-  const file = join(tmpdir(), `e2e-httpbin-${ns}.yaml`);
-  writeFileSync(file, manifest);
-  runOc(["apply", "-f", file], 60_000);
-  runOc(
-    ["-n", ns, "rollout", "status", "deployment/e2e-httpbin", "--timeout=180s"],
-    210_000,
-  );
 }
